move the special version function to its own file.  Since version is called by setup.py it must not import any "special" modules and cannot be within the csvpandas package.
